Fix race condition where display actions could execute out of order
Multithreading is so fun. This uses atomic boolean operations to force display locks be acquired in order -- even when thread::spawn() starts up at different times that doesn't match the order the notifier methods were called in. This previously resulted in scrolled message or display flashes executing in the wrong order sometimes.